// frontend/src/lib/openapi/model/externalPrincipalAccessCollection.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Teraslack API
 * Canonical HTTP API for Teraslack.
 * OpenAPI spec version: 1.0.0
 */
import type { ExternalPrincipalAccess } from './externalPrincipalAccess';

export interface ExternalPrincipalAccessCollection {
  items: ExternalPrincipalAccess[];
  next_cursor?: string;
}